debug

+++ springboot-QuerySystem/src/main/java/querysystem/model/dto/ProductDto.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class ProductDto {
 	public int Pid;
 	public String Ptype;//險種
-	public Double Pprice;//年繳
+	public Integer Pprice;//年繳
 	public Boolean Pstatus;// 上架中/停賣
 	public int Psales;//售出件數
 }
